Initial work on Splot class complete.
Splot is the the daemon and cache used to deliver the 'Arse That' blog
web pages. It sits between the scripts and memcached and
that's absolutely it.

It works out which order the articles go in and soon it'll render them in
GFM. It hasn't yet been optimised, whatsoever, since the data we'll be
working with initially will be very small. If the need arises it can be
very easily hacked.

Added:
	* Splot class
	* fetch, next and populate

TO DO:
	* Sinatra front end
	* Markdown
	* Cron job